tools/requirements.txt
kaleido
matplotlib
polars
plotly-express
